pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Page Analyzer"
authors = [
    { name = "Sroon Lady", email = "your@email.com" },
]
dependencies = [
    "Flask==3.0.2",
    "gunicorn==21.2.0",
    "python-dotenv==1.0.1",
    "psycopg2-binary==2.9.9",
    "requests==2.31.0",
    "validators==0.22.0",
    "blinker==1.9.0",
    "beautifulsoup4==4.12.3",
    "lxml==5.2.1",
]
requires-python = ">=3.9"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"